[api] improvements on neo4j client

+++ BookApp.API/Dtos/BookCatalogListDto.cs
@@ -1,6 +1,6 @@
 namespace BookApp.API.Dtos {
 
   public class BookCatalogListDto {
-    public int CatalogId { get; set; }
+    public long CatalogId { get; set; }
   }
 }
